refactor: 기능 중심 프로젝트 구조 수정  (#1)
