+++ src/features/savings/components/FilteredProducts.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ListRow } from 'tosslib';
-import { SavingsProduct } from '../types';
+import { SavingsProduct } from '../../../features/savings/types';
 import { default as ProductList } from './ProductList';
 
 type ProductListProps = {
